A lot of work to make picasso work again on Linux

+++ src/assets/shaders/attribute.h
@@ -25,8 +25,8 @@ class Attribute {
   std::string type_name;
   int location;
   GLenum type;
-  int count;
-  int type_size;
+  size_t count;
+  size_t type_size;
 };  // class Attribute
 
 }   // namespace shaders
